backend/requirements.txt
streamlit==1.41.1
pandas==2.2.3
psycopg[binary]
psycopg2-binary
plotly
requests
python-dotenv
openai
google-generativeai
sqlalchemy
friendli-client
matplotlib
boto3
scikit-learn
snowflake-sqlalchemy
snowflake-connector-python
streamlit
psycopg2-binary
pandas
python-dotenv
requests